+++ src/main/java/dev/prjbtrad001/app/bot/Status.java
@@ -3,6 +3,7 @@ package dev.prjbtrad001.app.bot;
 import jakarta.persistence.Embeddable;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Getter
@@ -15,17 +16,17 @@ public class Status {
 
   // Current state
   private boolean isLong;
-  private double purchasePrice;
+  private BigDecimal purchasePrice;
   private Instant purchaseTime;
-  private double quantity;
+  private BigDecimal quantity;
 
   // Status
-  private double lastPrice;
-  private double lastRsi;
-  private double lastSmaShort;
-  private double lastSmaLong;
-  private double actualSupport;
-  private double actualResistance;
-  private double lastVolume;
+  private BigDecimal lastPrice;
+  private BigDecimal lastRsi;
+  private BigDecimal lastSmaShort;
+  private BigDecimal lastSmaLong;
+  private BigDecimal actualSupport;
+  private BigDecimal actualResistance;
+  private BigDecimal lastVolume;
   
 }
